backend/pyproject.toml
[project]
name = "building-machinery-chatbot-backend"
version = "1.0.0"
description = "AI-powered chatbot backend for building machinery support"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    # FastAPI and web framework
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "sse-starlette",

    # Database drivers
    "motor",
    "pymongo",

    # Vector database (updated from deprecated pinecone-client)
    "pinecone",

    # AI and ML
    "openai",
    "pydantic-ai",
    "tiktoken",

    # Authentication and security
    "argon2-cffi",
    "python-jose[cryptography]",
    "itsdangerous",

    # Email
    "aiosmtplib",
    "jinja2",

    # HTTP client
    "httpx",

    # Environment and configuration
    "python-dotenv",
    "pydantic-settings",
    "pydantic[email]",

    # Monitoring and logging
    "prometheus-fastapi-instrumentator",
    "sentry-sdk[fastapi]",
    "structlog",

    # Document processing
    "docling",
    "PyPDF2",
    "python-docx",
    "openpyxl",
    "python-pptx",
    "Pillow",

    # Utilities
    "python-dateutil",
]

[project.optional-dependencies]
dev = [
    # Code quality
    "black",
    "flake8",
    "mypy",
    "ruff",
]

test = [
    # Testing
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-xdist",
    "pytest-mock",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "black",
    "flake8",
    "mypy",
    "ruff",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-xdist",
    "pytest-mock",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "motor.*",
    "pymongo.*",
    "pinecone.*",
    "docling.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
